refactor(ed25519:secp256k1): Remove node-only dependencies (#21)
# Motivation
We wish to run `ecdsa/secp256k1.ts` in a browser, but it currently
depends on two node dependencies: `crypto` and `assert`.

# Changes
- Replace `crypto.createHmac` with the same function provided by the npm
module `create-hmac` which uses the node crypto when available, else the
browser equivalent.
- Replace the call to `assert` with an explicit check & exception.

# Tests
For nodejs, CI suffices.

For how to run a CI check that no node-specific APIs are used in the
library (not the cli), suggestions are welcome.