feat(typescript): Basic eslint config (#33)
# Motivation
The code will need a typescript linter.

There are a lot of lints for basic stylistic issues such as node-style
camelcase variables, so this starts with a very simple eslint config.
Refinements should be added later, to limit PR size.

# Changes
- Add eslint configuration.

Note: This PR does not include the eslint fixes and does not apply
eslint in CI.

# Tests
Eslint runs locally and complains about many things.