feat(cli): Derive chain fusion signer btc addresses (#36)
# Motivation
A common use case is deriving chain fusion signer addresses. IN this
case the BTC addresses (different for each network).

# Changes

- Install library
[bitcoinjs-lib](https://www.npmjs.com/package/bitcoinjs-lib): it will be
useful to simplify the calculation from public key to BTC address (note
that in the signer, we rely on a
[bitcoin](https://crates.io/crates/bitcoin) library too).
- Add a cli option to derive BTC addresses, defaulting to the production
chain fusion signer parameters: it is very similar to how the ETH
address is derived, with the addition of the network input.

# Note

There is quite a lot of repeated code, but I would prefer to refactor
and optimize in another PR.

# Tests
A unit test is included that checks the cli output against the
production chain fusion signer.

For a manual test, run this with your production Oisy principal:
```
npx ic-pub-key signer btc address --user YOUR_PRINCIPAL --network mainnet
```
and compare the output with your OISY BTC mainnet address (same for
testnet and regtest).