refactor: Make `schnorrEd25519Derive` service return the response (#90)
# Motivation

Since service `schnorrEd25519Derive` is going to be re-used by any
number of consumer, it would be preferably that it returns an object
with the response, instead of printing to console.

# Changes

- Change the return type of `schnorrEd25519Derive` to be a structure
with request and response (similar to other services in the library).
- Adapt the CLI methods to print to console as before.
- Change all variable related to `schnorrEd25519Derive` to use
camelCase.

# Tests

Current tests should work as expected. Nothing changed.